nixos/grub-install: execute prepare commands earlier
Previously the extraPrepareConfig commands where exectued after the grub
entries were generated. With this new location we can run commands
before the initrd appender scripts is executed. This is especially
helpful with the recent change to the sshd host keys of openssh in the
initrd. In the old setup you did not have to provision host keys (and
you might not even have cared about them) but with the new setup you
must always provide host keys even if you do not care about them.

In my personal setup I have everything encrypted except that initrd. The
initrd is basically public knowledge as anyone (on the hosting provider)
will be able to read that key from the initrd.