python3Packages.attrs: separate test run into a separate tests attribute.
This resolves a circular dependency issue between pytest and attrs, by
instead building a separate output that _just_ contains the tests from
the original package, which is then consumed by a separate tests
derivation.

The downside of this approach is that these tests will not be run on
Hydra. They're not being run on Hydra at the moment, either, since
doCheck is false.